Implemented reusing groupme tabs
Note: currently changes the url of the tab to be reused.
Not always the best idea because it reloads the page to show the
appropriate group (even if that group is already showing on the page).

I'm thinking of just leaving the url as is and finding some way to get
the appropriate chat to display (maybe there is something in the API?)